wpt_babel/content-security-policy/script-src/script-src-strict_dynamic_eval.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>
   Scripts injected via `eval` are not allowed with `strict-dynamic` without `unsafe-eval`.
  </title>
  <script nonce="dummy" src="/resources/testharness.js">
  </script>
  <script nonce="dummy" src="/resources/testharnessreport.js">
  </script>
  <!-- CSP served: script-src 'strict-dynamic' 'nonce-dummy' -->
 </head>
 <body>
  <h1>
   Scripts injected via `eval` are not allowed with `strict-dynamic` without `unsafe-eval`.
  </h1>
  <div id="log">
  </div>
  <script nonce="dummy">
   "use strict";

var evalScriptRan = false;
async_test(function (t) {
  window.addEventListener('securitypolicyviolation', t.step_func_done(function (e) {
    assert_false(evalScriptRan);
    assert_equals(e.effectiveDirective, 'script-src');
    assert_equals(e.blockedURI, 'eval');
  }));
  assert_throws_js(Error, function () {
    try {
      eval("evalScriptRan = true;");
    } catch (e) {
      throw new Error();
    }
  });
}, "Script injected via `eval` is not allowed with `strict-dynamic` without `unsafe-eval`.");
  </script>
 </body>
</html>
